Modulo salas y sillas, faltando controllers

+++ src/main/java/CinePacho/demo/snacks/dto/request/SnackRequest.java
@@ -2,7 +2,7 @@ package CinePacho.demo.snacks.dto.request;
 
 import jakarta.validation.constraints.*;
 import lombok.*;
-
+import java.math.BigDecimal;
  
 @Getter
 @Setter
@@ -21,7 +21,7 @@ public class SnackRequest {
     @NotNull(message = "El precio es obligatorio")
     @DecimalMin(value = "0.0", inclusive = false, message = "El precio debe ser mayor a 0")
     @Digits(integer = 8, fraction = 2, message = "El precio debe tener máximo 8 dígitos enteros y 2 decimales")
-    private double price;
+    private BigDecimal price;
 
     @NotNull(message = "La cantidad es obligatoria")
     @Min(value = 0, message = "La cantidad no puede ser negativa")
